requirements.txt
asn1crypto==0.23.0
backports.ssl-match-hostname==3.5.0.1
cffi==1.11.2
cryptography==2.1.3
disco-py==0.0.11
enum34==1.1.6
erlpack==0.3.2
gevent==1.2.2
greenlet==0.4.12
holster==1.0.16
idna==2.6
ipaddress==1.0.18
pycparser==2.18
pyOpenSSL==17.3.0
requests==2.13.0
six==1.10.0
ujson==1.35
websocket-client==0.40.0
